feat: Remove streaming.parsePrftBox config
