+++ lib/media/media_source_engine.js
@@ -22,6 +22,7 @@ goog.require('shaka.util.BufferUtils');
 goog.require('shaka.util.Destroyer');
 goog.require('shaka.util.Error');
 goog.require('shaka.util.EventManager');
+goog.require('shaka.util.FakeEvent');
 goog.require('shaka.util.Functional');
 goog.require('shaka.util.IDestroyable');
 goog.require('shaka.util.Id3Utils');
@@ -166,6 +167,9 @@ shaka.media.MediaSourceEngine = class {
 
     /** @private {!shaka.util.PublicPromise.<number>} */
     this.audioCompensation_ = new shaka.util.PublicPromise();
+
+    /** @private {boolean} */
+    this.parsedPrftEventRaised_ = false;
   }
 
   /**
@@ -498,6 +502,12 @@ shaka.media.MediaSourceEngine = class {
         !this.ignoreManifestTimestampsInSegmentsMode_;
 
     this.tsParser_ = null;
+    this.firstVideoTimestamp_ = null;
+    this.firstVideoReferenceStartTime_ = null;
+    this.firstAudioTimestamp_ = null;
+    this.firstAudioReferenceStartTime_ = null;
+    this.audioCompensation_ = new shaka.util.PublicPromise();
+    this.parsedPrftEventRaised_ = false;
 
     for (const contentType of streamsByType.keys()) {
       const stream = streamsByType.get(contentType);
@@ -838,8 +848,11 @@ shaka.media.MediaSourceEngine = class {
         const Mp4Parser = shaka.util.Mp4Parser;
         let startTime = 0;
         let parsedMedia = false;
-        new Mp4Parser()
-            .box('moof', Mp4Parser.children)
+        const parser = new Mp4Parser();
+        if (!this.parsedPrftEventRaised_) {
+          parser.fullBox('prft', (box) => this.parsePrft_(timescale, box));
+        }
+        parser.box('moof', Mp4Parser.children)
             .box('traf', Mp4Parser.children)
             .fullBox('tfdt', (box) => {
               goog.asserts.assert(
@@ -874,6 +887,48 @@ shaka.media.MediaSourceEngine = class {
       }
     }
     return timestamp;
+  }
+
+  /**
+   * Parse PRFT box.
+   * @param {number} timescale
+   * @param {!shaka.extern.ParsedBox} box
+   * @private
+   */
+  parsePrft_(timescale, box) {
+    goog.asserts.assert(
+        box.version == 0 || box.version == 1,
+        'PRFT version can only be 0 or 1');
+    const parsed = shaka.util.Mp4BoxParsers.parsePRFTInaccurate(
+        box.reader, box.version);
+
+    const wallClockTime = this.convertNtp_(parsed.ntpTimestamp);
+    const programStartDate = new Date(wallClockTime -
+      (parsed.mediaTime / timescale) * 1000);
+    const prftInfo = {
+      wallClockTime,
+      programStartDate,
+    };
+
+    const eventName = shaka.util.FakeEvent.EventName.Prft;
+    const data = (new Map()).set('detail', prftInfo);
+    const event = new shaka.util.FakeEvent(
+        eventName, data);
+    this.playerInterface_.onEvent(event);
+    this.parsedPrftEventRaised_ = true;
+  }
+
+
+  /**
+   * Convert Ntp ntpTimeStamp to UTC Time
+   *
+   * @param {number} ntpTimeStamp
+   * @return {number} utcTime
+   * @private
+   */
+  convertNtp_(ntpTimeStamp) {
+    const start = new Date(Date.UTC(1900, 0, 1, 0, 0, 0));
+    return new Date(start.getTime() + ntpTimeStamp).getTime();
   }
 
   /**
@@ -2234,7 +2289,8 @@ shaka.media.MediaSourceEngine.SourceBufferMode_ = {
 /**
  * @typedef {{
  *   getKeySystem: function():?string,
- *   onMetadata: function(!Array<shaka.extern.ID3Metadata>, number, ?number)
+ *   onMetadata: function(!Array<shaka.extern.ID3Metadata>, number, ?number),
+ *   onEvent: function(!Event)
  * }}
  *
  * @summary Player interface
@@ -2243,5 +2299,7 @@ shaka.media.MediaSourceEngine.SourceBufferMode_ = {
  * @property {function(
  *     !Array<shaka.extern.ID3Metadata>, number, ?number)} onMetadata
  *   Callback to use when metadata arrives.
+ * @property {function(!Event)} onEvent
+ *   Called when an event occurs that should be sent to the app.
  */
 shaka.media.MediaSourceEngine.PlayerInterface;
