+++ lib/media/streaming_engine.js
@@ -93,9 +93,6 @@ shaka.media.StreamingEngine = class {
     /** @private {number} */
     this.textStreamSequenceId_ = 0;
 
-    /** @private {boolean} */
-    this.parsedPrftEventRaised_ = false;
-
     /**
      * Maps a content type, e.g., 'audio', 'video', or 'text', to a MediaState.
      *
@@ -2245,8 +2242,6 @@ shaka.media.StreamingEngine = class {
     const hasEmsg = ((stream.emsgSchemeIdUris != null &&
       stream.emsgSchemeIdUris.length > 0) ||
       this.config_.dispatchAllEmsgBoxes);
-    const shouldParsePrftBox =
-      (this.config_.parsePrftBox && !this.parsedPrftEventRaised_);
 
     const isMP4 = stream.mimeType == 'video/mp4' ||
             stream.mimeType == 'audio/mp4';
@@ -2259,7 +2254,7 @@ shaka.media.StreamingEngine = class {
         this.manifest_.type == shaka.media.ManifestParser.DASH &&
         this.config_.shouldFixTimestampOffset;
 
-    if (hasEmsg || shouldParsePrftBox || shouldFixTimestampOffset) {
+    if (hasEmsg || shouldFixTimestampOffset) {
       parser = new shaka.util.Mp4Parser();
     }
 
@@ -2269,14 +2264,6 @@ shaka.media.StreamingEngine = class {
               'emsg',
               (box) => this.parseEMSG_(
                   reference, stream.emsgSchemeIdUris, box));
-    }
-
-    if (shouldParsePrftBox) {
-      parser
-          .fullBox(
-              'prft',
-              (box) => this.parsePrft_(
-                  reference, box));
     }
 
     if (shouldFixTimestampOffset) {
@@ -2317,7 +2304,7 @@ shaka.media.StreamingEngine = class {
           });
     }
 
-    if (hasEmsg || shouldParsePrftBox || shouldFixTimestampOffset) {
+    if (hasEmsg || shouldFixTimestampOffset) {
       parser.parse(segment, /* partialOkay= */ false, isChunkedData);
     }
 
@@ -2465,52 +2452,6 @@ shaka.media.StreamingEngine = class {
         }
       }
     }
-  }
-
-  /**
-   * Parse PRFT box.
-   * @param {!shaka.media.SegmentReference} reference
-   * @param {!shaka.extern.ParsedBox} box
-   * @private
-   */
-  parsePrft_(reference, box) {
-    if (this.parsedPrftEventRaised_ ||
-      !reference.initSegmentReference.timescale) {
-      return;
-    }
-    goog.asserts.assert(
-        box.version == 0 || box.version == 1,
-        'PRFT version can only be 0 or 1');
-    const parsed = shaka.util.Mp4BoxParsers.parsePRFTInaccurate(
-        box.reader, box.version);
-
-    const timescale = reference.initSegmentReference.timescale;
-    const wallClockTime = this.convertNtp(parsed.ntpTimestamp);
-    const programStartDate = new Date(wallClockTime -
-      (parsed.mediaTime / timescale) * 1000);
-    const prftInfo = {
-      wallClockTime,
-      programStartDate,
-    };
-
-    const eventName = shaka.util.FakeEvent.EventName.Prft;
-    const data = (new Map()).set('detail', prftInfo);
-    const event = new shaka.util.FakeEvent(
-        eventName, data);
-    this.playerInterface_.onEvent(event);
-    this.parsedPrftEventRaised_ = true;
-  }
-
-
-  /**
-   * Convert Ntp ntpTimeStamp to UTC Time
-   *
-   * @param {number} ntpTimeStamp
-   * @return {number} utcTime
-   */
-  convertNtp(ntpTimeStamp) {
-    const start = new Date(Date.UTC(1900, 0, 1, 0, 0, 0));
-    return new Date(start.getTime() + ntpTimeStamp).getTime();
   }
 
   /**
