+++ lib/player.js
@@ -2405,6 +2405,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
           onMetadata: (metadata, offset, endTime) => {
             this.processTimedMetadataMediaSrc_(metadata, offset, endTime);
           },
+          onEvent: (event) => this.dispatchEvent(event),
         },
         this.lcevcDec_);
     mediaSourceEngine.configure(this.config_.mediaSource);
