+++ lib/util/player_configuration.js
@@ -233,7 +233,6 @@ shaka.util.PlayerConfiguration = class {
       dispatchAllEmsgBoxes: false,
       observeQualityChanges: false,
       maxDisabledTime: 30,
-      parsePrftBox: false,
       // When low latency streaming is enabled, segmentPrefetchLimit will
       // default to 2 if not specified.
       segmentPrefetchLimit: 0,
