+++ test/media/streaming_engine_integration.js
@@ -73,6 +73,7 @@ describe('StreamingEngine', () => {
         {
           getKeySystem: () => null,
           onMetadata: () => {},
+          onEvent: () => {},
         });
     const mediaSourceConfig =
         shaka.util.PlayerConfiguration.createDefault().mediaSource;
